Split utils into separate files
