+++ beta/src/components/MDX/Sandpack/Preview.tsx
@@ -8,8 +8,11 @@ import {useSandpack, LoadingOverlay} from '@codesandbox/sandpack-react';
 import cn from 'classnames';
 
 import {Error} from './Error';
-import {computeViewportSize, generateRandomId} from './utils';
-import type {LintDiagnostic} from './utils';
+import {computeViewportSize} from './computeViewportSize';
+import type {LintDiagnostic} from './useSandpackLint';
+
+const generateRandomId = (): string =>
+  Math.floor(Math.random() * 10000).toString();
 
 type CustomPreviewProps = {
   className?: string;
